Add attribution to plugins list
