+++ src/components/backstage/plugins/index.js
@@ -1,0 +1,3 @@
+export { default as Attribution } from './Attribution';
+export { default as Logo } from './Logo';
+export { default as ListItem } from './ListItem';
