+++ src/pages/backstage/plugins.js
@@ -57,6 +57,11 @@ export const pageQuery = graphql`
           name
           humanName
 
+          attribution {
+            text
+            href
+          }
+
           childrenLogoImage {
             childImageSharp {
               fixed(width: 200, grayscale: false) {
